% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_extract.R
\name{ehr_extract}
\alias{ehr_extract}
\title{Extract EHR records and values}
\usage{
ehr_extract(ehr_data, read_codes, values = NULL)
}
\arguments{
\item{ehr_data}{Data table with UK Biobank clinical event data i.e.
\code{gp_clinical.txt}.}

\item{read_codes}{Extract records matching these codes. Data table/frame
with \code{read_2} (Read v2) and \code{read_3} (CTV3) columns.}

\item{values}{Extract values from EHR records. Replaces the \code{value1},
\code{value2}, \code{value3} columns with a single \code{value} column and
a \code{unit} column. Default is currently TRUE but will change to FALSE
in a future release.}
}
\value{
Filtered EHR data with \code{value} and \code{unit} columns if
  \code{values} is TRUE.
}
\description{
Extracts records/values from linked EHR data. If \code{values}
  is TRUE, values are extracted the \code{value1} field with the exception of
  data provider 2 where values are extracted from \code{value2} if
  \code{value1} is empty. Units are taken from \code{value3} for data
  provider 2 (otherwise units are unavailable). \code{NA}, zero and duplicate
  values are dropped.
}
