% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_map_codes.R
\name{ehr_map_codes}
\alias{ehr_map_codes}
\title{Map EHR codes to another terminology}
\usage{
ehr_map_codes(codes, from = "read2", to = "ctv3", overwrite = FALSE)
}
\arguments{
\item{codes}{Vector of codes to map.}

\item{from}{Terminology to map from. Currently supports Read v2 codes only.}

\item{to}{Terminology to map to. Currently supports CTV3 only.}

\item{overwrite}{Overwrite existing mapping dictionary (default FALSE).}
}
\value{
Data table with mapped codes and description.
}
\description{
Map EHR codes from one terminology to another using UK Biobank mapping
dictionaries. Currently supports mapping Read v2 to CTV 3 only.
}
