% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding.R
\name{get_coding}
\alias{get_coding}
\title{Get UK Biobank coding}
\usage{
get_coding(id, overwrite = FALSE)
}
\arguments{
\item{id}{ID for coding file. Obtain from UK Biobank Showcase at \href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}.}

\item{overwrite}{Overwrite existing file (default FALSE).}
}
\value{
Data table with contents of coding file.
}
\description{
File is downloaded to root directory if it is unavailable locally.
}
