% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr_concept.R
\name{get_hdr_concept}
\alias{get_hdr_concept}
\title{Get concept from HDR UK phenotype library}
\usage{
get_hdr_concept(id, version_id = NULL, api = NULL, read = FALSE)
}
\arguments{
\item{id}{Concept ID.}

\item{version_id}{Optional concept version ID. Otherwise latest version of
concept is returned.}

\item{api}{Optional HttpClient to connect to phenotype library (default uses
\href{https://phenotypes.healthdatagateway.org/}{https://phenotypes.healthdatagateway.org/}).}

\item{read}{Set to TRUE if concept contains Read v2 codes. Read v2 codes of
length less than 5 are padded with a "." e.g. "C10" becomes "C10..". Codes
are mapped to CTV3 using UK Biobank mapping dictionaries.}
}
\value{
Data table with phenotype codes. \code{read_2} (Read v2) and
  \code{read_3} (CTV3) columns are added if \code{read} is TRUE.
}
\description{
Queries HDR UK phenotype library for a given concept. Read v2 codes are
mapped to CTV3 if the \code{read} argument is TRUE.
}
