% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr_phenotype.R
\name{get_hdr_phenotype}
\alias{get_hdr_phenotype}
\title{Get phenotype from HDR UK phenotype library}
\usage{
get_hdr_phenotype(id, version_id = NULL, api = NULL)
}
\arguments{
\item{id}{Phenotype ID.}

\item{version_id}{Optional phenotype version ID. Otherwise latest version of
phenotype is returned.}

\item{api}{Optional HttpClient to connect to phenotype library (default uses
\href{https://phenotypes.healthdatagateway.org/}{https://phenotypes.healthdatagateway.org/}).}
}
\value{
Data table with phenotype codes.
}
\description{
Queries HDR UK phenotype library for a given phenotype.
}
