% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visit}
\alias{visit}
\title{Example UK Biobank visit data set.}
\format{
Data table with columns as set out below.
}
\usage{
visit
}
\description{
Synthetic UK Biobank visit data for function testing and demonstration.
}
\details{
Data is provided for 100 synthetic participants with two instances (visits)
for 40\% of participants. Fields provided are \code{50} (height), \code{53}
(visit date) and \code{21002} (weight).
}
\keyword{datasets}
