% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_cancer.R
\name{visit_cancer}
\alias{visit_cancer}
\title{Extract self-reported cancer history}
\usage{
visit_cancer(visit_data)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data. Must include fields
\code{20001} and \code{20006}.}
}
\value{
Data table with the following columns: \describe{
  \item{eid}{UK Biobank identifier.}
  \item{date}{Reported date of diagnosis.}
  \item{condition}{Cancer type coded as in
  \href{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=3}{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=3}.}
  \item{desc}{Description of cancer type (added if coding data can be downloaded).}
  \item{reported}{Visit date at which the cancer was self-reported.}
}
}
\description{
Extracts self-reported cancer history in a "long" format that is
  easier to work with than "wide" as provided by UK Biobank (NOTE: watch for
  type coercion of different data types). Wrapper function for
  \code{visit_mult_array()}.
}
