% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_conditions.R
\name{visit_conditions}
\alias{visit_conditions}
\title{Extract self-reported non-cancer medical history}
\usage{
visit_conditions(visit_data)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data. Must include fields
\code{20002} and \code{20008}.}
}
\value{
Data table with the following columns: \describe{
  \item{eid}{UK Biobank identifier.}
  \item{date}{Reported date of diagnosis.}
  \item{condition}{Health condition coded as in
  \href{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=6}{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=6}.}
  \item{desc}{Description of health condition (added if coding data can be downloaded).}
  \item{reported}{Visit date at which the condition was self-reported.}
}
}
\description{
Extracts self-reported non-cancer medical history in a "long"
  format that is easier to work with than "wide" as provided by UK Biobank
  (NOTE: watch for type coercion of different data types). Wrapper
  function for \code{visit_mult_array()}.
}
