% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_extract.R
\name{visit_extract}
\alias{visit_extract}
\title{Extract field data from UK Biobank visit data}
\usage{
visit_extract(visit_data, fields, format = NULL)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data.}

\item{fields}{Vector of fields to extract e.g. \code{50} or
\code{c(50, 21002)}. Field name will be identified from UK Biobank schema.
Alternatively, field names can be set using a named vector e.g.
\code{c("height" = 50, "weight" = 21002)}.}

\item{format}{Format of output table (\code{raw} or \code{source}). Default
is currently raw but will change to source in a future release.}
}
\value{
Data table with values of all instances/arrays for each field in
  "long" format. The following columns are provided: \describe{
    \item{eid}{UK Biobank identifier.}
    \item{date}{Visit date.}
    \item{field/variable}{Field name (see below).}
    \item{array}{Provided if any fields have multiple arrays (more than one
    value recorded on the same date e.g repeated blood pressure).}
    \item{value}{Value recorded.}
  }
  If \code{format} = "source", an additional column \code{source} = "ukbb" is
  added to indicate data was recorded by UK Biobank and the field column is
  renamed \code{variable}. This will be the default in a future release. Use
  \code{format} = "raw" to keep current format.
}
\description{
Extracts all instances/arrays of data for a UK Biobank
  field(s) in clean "long" format (NOTE: watch for type coercion of
  different data types). See \href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}
  to identify field codes. Wrapper for \code{visit_fields()} which extracts
  raw field data.
}
\examples{
\dontrun{
# Load data
data_path <- ""  # add path to your data
visit_data <- fread(data_path)

# Extract a field
visit_extract(visit_data, 50)

# Extract multiple fields
visit_extract(visit_data, c(50, 21002))

# Manually specify a field name
visit_extract(visit_data, c("height" = 50, 21002))
}

}
