% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_family_history.R
\name{visit_family_history}
\alias{visit_family_history}
\title{Determine family history of a specified condition}
\usage{
visit_family_history(
  visit_data,
  fields,
  condition,
  collapse = TRUE,
  name = NULL
)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data.}

\item{fields}{Vector of family history fields to extract e.g. one or more
from \code{20107} (father), \code{20110} (mother) or \code{20111}
(sibling).}

\item{condition}{Code for condition. Only a single condition at a time is
currently supported. See \href{https://biobank.ndph.ox.ac.uk/showcase/coding.cgi?id=1010}{https://biobank.ndph.ox.ac.uk/showcase/coding.cgi?id=1010}
to identify condition codes.}

\item{collapse}{Summarise results across all visit dates (default \code{TRUE}).
If \code{FALSE}, the presence of the condition is provided at each visit
date in the output table.}

\item{name}{Optional column name for condition (default \code{history}).}
}
\value{
Data table with \code{TRUE} (condition was reported), \code{FALSE}
  (condition was not reported) or \code{NA} (unknown/no response).
}
\description{
Determines presence of a specified condition in the self-reported
  family history data. If multiple history fields are provided (e.g. history
  of mother and father), presence of the condition in either field determines
  a positive family history.
}
\examples{
\dontrun{
# Load data
data_path <- ""  # add path to your data
visit <- fread(data_path)

# Extract history for father
visit_family_history(visit, 20107, 9)

# Extract history for father, mother and siblings
visit_family_history(visit, c(20107, 20110, 20111), 9)

# Name column in output
visit_family_history(visit, c(20107, 20110, 20111), 9, name = "diabetes")

# Get history reported at each visit date
visit_family_history(visit, c(20107, 20110, 20111), 9, collapse = FALSE)
}

}
