% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fields.R
\name{visit_fields}
\alias{visit_fields}
\title{Extract raw field data from UK Biobank visit data}
\usage{
visit_fields(visit_data, fields, format = "wide")
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data.}

\item{fields}{Vector of fields to extract e.g. \code{50} or
\code{c(50, 21002)}.}

\item{format}{Format of output table i.e. "wide" or "long" (default "wide").
NOTE: watch for type coercion if use long format.}
}
\value{
Data table with all instances/arrays for each field.
}
\description{
Extracts all instances/arrays of data for a UK Biobank
  field(s). See \href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}
  to identify field codes.
}
\examples{
\dontrun{
# Load data
data_path <- ""  # add path to your data
visit_data <- fread(data_path)

# Extract a field
visit_fields(visit_data, 50)

# Extract multiple fields
visit_fields(visit_data, c(50, 21002))

# Extract multiple fields in long format
visit_fields(visit_data, c(50, 21002), format = "long")
}

}
