% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_mult_array.R
\name{visit_mult_array}
\alias{visit_mult_array}
\title{Extract data from two UK Biobank fields jointly}
\usage{
visit_mult_array(visit_data, fields)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data.}

\item{fields}{Vector of fields to extract e.g. \code{c(50, 21002)}. Must
be length two. Field name will be identified from UK Biobank schema.
Alternatively, field names can be set using a named vector e.g.
\code{c("height" = 50, "weight" = 21002)}.}
}
\value{
Data table with \code{eid}, \code{reported}, and columns
  corresponding to the \code{fields} argument. \code{reported} is the date
  corresponding to the field instance e.g. the UK Biobank visit at which the
  data was collected. Each row shows the data for an array.
}
\description{
Some fields relate to each other e.g. self-reported medical
history where field 20002 contains the disclosed conditions and 20008 the
date of diagnosis. The date in array \code{i} of 20008 corresponds to the
condition in array \code{i} of 20002. \code{visit_mult_array()} jointly
extracts such fields in a "long" format that is easier to work with than
"wide" as provided by UK Biobank (NOTE: watch for type coercion of
different data types).
}
