% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_subset.R
\name{visit_subset}
\alias{visit_subset}
\title{Subset fields from UK Biobank visit data}
\usage{
visit_subset(data_path, fields, save = NULL)
}
\arguments{
\item{data_path}{Path to raw UK Biobank data unpacked using \code{ukbunpack}
utility.}

\item{fields}{Vector of fields to extract e.g. \code{50} or
\code{c(50, 21002)}.}

\item{save}{Optional path to save output.}
}
\value{
Data table with all instances/arrays for each field. Note the date
  field (53) is always returned.
}
\description{
Loads UK Biobank data and subsets required fields. Use to avoid
  loading full data each time. See \href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}
  to identify field codes.
}
