% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_extract.R
\name{ehr_extract}
\alias{ehr_extract}
\title{Extract observations/test results from raw EHR data}
\usage{
ehr_extract(ehr_data, read_codes)
}
\arguments{
\item{ehr_data}{Data table/frame with UK Biobank clinical event data i.e.
\code{gp_clinical.txt}.}

\item{read_codes}{Values are extracted from these codes. Data table/frame
with Read v2 (\code{read_2}) and CTV3 (\code{read_3}) columns.}
}
\value{
Data table with \code{value} and \code{unit} columns.
}
\description{
Extracts values from linked EHR data. Data is extracted from
  the \code{value1} field with the exception of data provider 2 where values
  are extracted from \code{value2} if \code{value1} is empty. Units are taken
  from \code{value3} for data provider 2 (otherwise units are unavailable).
  \code{NA}, zero and duplicate values are dropped.
}
