% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding.R
\name{get_coding}
\alias{get_coding}
\title{Get UK Biobank coding}
\usage{
get_coding(id, overwrite = FALSE)
}
\arguments{
\item{id}{ID for coding file. Obtain this from the data dictionary for your application, or directly from the Data Showcase (\href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}), by looking up the relevant data field.}

\item{overwrite}{Overwrite existing file (default FALSE).}
}
\value{
Data table with contents of coding file.
}
\description{
Returns the UK Biobank coding file for the supplied \code{id}. The file is downloaded if it is unavailable locally.
}
