% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_cancer.R
\name{visit_cancer}
\alias{visit_cancer}
\title{Extract self-reported cancer history}
\usage{
visit_cancer(visit_data)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data. Must include fields
\code{20001} and \code{20006}.}
}
\value{
Data table with \code{eid}, \code{date}, \code{condition} and
  \code{reported} columns. \code{date} is the reported date of diagnosis and
  \code{reported} is the date the cancer was self-reported to UK Biobank.
  \code{condition} is the cancer type coded as in
  \href{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=3}{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=3}.
}
\description{
Extracts self-reported cancer history in a "long" format that is
  easier to work with than "wide" as provided by UK Biobank.
}
