% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_extract.R
\name{visit_extract}
\alias{visit_extract}
\title{Extract fields from UK Biobank visit data}
\usage{
visit_extract(visit_data, fields)
}
\arguments{
\item{visit_data}{Data frame/table with UK Biobank data.}

\item{fields}{Vector of fields to extract e.g. \code{50} or
\code{c(50, 21002)}. Field name will be identified from UK Biobank schema.
Alternatively, field names can be set using a named vector e.g.
\code{c("height" = 50, "weight" = 21002)}.}
}
\value{
Data table with values of all instances/arrays for each field in
  "long" format i.e. \code{eid}, \code{date} of visit, \code{field} name and
  \code{value} recorded by UK Biobank.
}
\description{
Extracts all instances/arrays of data for a UK Biobank
  field(s). See \href{https://biobank.ndph.ox.ac.uk/showcase/}{https://biobank.ndph.ox.ac.uk/showcase/}
  to identify field codes.
}
\examples{
\dontrun{
# Load data
data_path <- ""  # add path to your data
visit <- fread(data_path)

# Extract a field
visit_extract(visit, 50)

# Extract multiple fields
visit_extract(visit, c(50, 21002))

# Manually specify a field name
visit_extract(visit, c("height" = 50, 21002))
}

}
